# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Argus SBOM Guard",
    "version": "0.0.8-beta"
  },
  "paths": {
...
    "/api/v1/sboms/upload": {
      "post": {
        "tags": [
          "sboms"
        ],
        "summary": "Upload Sbom",
        "operationId": "upload_sbom_api_v1_sboms_upload_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_sbom_api_v1_sboms_upload_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SBOMUploadResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid API key"
          },
          "404": {
            "description": "Resource not found"
          },
          "400": {
            "description": "Malformed request body"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ]
      }
    },
    "/api/v1/sboms/{sbom_id}/download": {
      "get": {
        "tags": [
          "sboms"
        ],
        "summary": "Download Sbom",
        "operationId": "download_sbom_api_v1_sboms__sbom_id__download_get",
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "sbom_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Sbom Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "401": {
            "description": "Missing or invalid API key"
          },
          "404": {
            "description": "Resource not found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/sboms/{sbom_id}": {
      "get": {
        "tags": [
          "sboms"
        ],
        "summary": "Get Sbom",
        "operationId": "get_sbom_api_v1_sboms__sbom_id__get",
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "sbom_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Sbom Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SBOMDetailResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid API key"
          },
          "404": {
            "description": "Resource not found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "sboms"
        ],
        "summary": "Delete Sbom",
        "operationId": "delete_sbom_api_v1_sboms__sbom_id__delete",
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "sbom_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Sbom Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "401": {
            "description": "Missing or invalid API key"
          },
          "404": {
            "description": "Resource not found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/sboms/{sbom_id}/diff/{other_id}": {
      "get": {
        "tags": [
          "sboms"
        ],
        "summary": "Diff Sboms",
        "operationId": "diff_sboms_api_v1_sboms__sbom_id__diff__other_id__get",
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "sbom_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Sbom Id"
            }
          },
          {
            "name": "other_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Other Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SBOMDiffResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid API key"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/services": {
      "get": {
        "tags": [
          "services"
        ],
        "summary": "List Services",
        "operationId": "list_services_api_v1_services_get",
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "project_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Project Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ServiceResponse"
                  },
                  "title": "Response List Services Api V1 Services Get"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid API key"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/services/{service_id}": {
      "delete": {
        "tags": [
          "services"
        ],
        "summary": "Delete Service",
        "operationId": "delete_service_api_v1_services__service_id__delete",
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "service_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Service Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "401": {
            "description": "Missing or invalid API key"
          },
          "404": {
            "description": "Resource not found"
          },
          "409": {
            "description": "Resource already exists"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_upload_sbom_api_v1_sboms_upload_post": {
        "properties": {
          "project_id": {
            "type": "string",
            "title": "Project Id"
          },
          "slug": {
            "type": "string",
            "title": "Slug"
          },
          "version": {
            "type": "string",
            "title": "Version"
          },
          "service_name": {
            "type": "string",
            "title": "Service Name"
          },
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_upload_sbom_api_v1_sboms_upload_post"
      },
      "DependencyResponse": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "version": {
            "type": "string",
            "title": "Version"
          },
          "purl": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Purl"
          },
          "type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Type"
          },
          "license": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "License"
          },
          "is_direct": {
            "type": "boolean",
            "title": "Is Direct"
          }
        },
        "type": "object",
        "required": [
          "name",
          "version",
          "purl",
          "type",
          "license",
          "is_direct"
        ],
        "title": "DependencyResponse"
      },
      "DiffItemResponse": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "version": {
            "type": "string",
            "title": "Version"
          }
        },
        "type": "object",
        "required": [
          "name",
          "version"
        ],
        "title": "DiffItemResponse"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "SBOMDetailResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "project_id": {
            "type": "string",
            "format": "uuid",
            "title": "Project Id"
          },
          "version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Version"
          },
          "format": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Format"
          },
          "sha256": {
            "type": "string",
            "title": "Sha256"
          },
          "dependency_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dependency Count"
          },
          "created_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          },
          "dependencies": {
            "items": {
              "$ref": "#/components/schemas/DependencyResponse"
            },
            "type": "array",
            "title": "Dependencies"
          },
          "vulnerabilities": {
            "items": {
              "$ref": "#/components/schemas/VulnerabilityBriefResponse"
            },
            "type": "array",
            "title": "Vulnerabilities"
          }
        },
        "type": "object",
        "required": [
          "id",
          "project_id",
          "version",
          "format",
          "sha256",
          "dependency_count",
          "created_at",
          "dependencies",
          "vulnerabilities"
        ],
        "title": "SBOMDetailResponse"
      },
      "SBOMDiffResponse": {
        "properties": {
          "added": {
            "items": {
              "$ref": "#/components/schemas/DiffItemResponse"
            },
            "type": "array",
            "title": "Added"
          },
          "removed": {
            "items": {
              "$ref": "#/components/schemas/DiffItemResponse"
            },
            "type": "array",
            "title": "Removed"
          },
          "changed": {
            "items": {
              "$ref": "#/components/schemas/VersionChangeResponse"
            },
            "type": "array",
            "title": "Changed"
          }
        },
        "type": "object",
        "required": [
          "added",
          "removed",
          "changed"
        ],
        "title": "SBOMDiffResponse"
      },
      "SBOMUploadResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "format": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Format"
          },
          "dependency_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dependency Count"
          },
          "sha256": {
            "type": "string",
            "title": "Sha256"
          }
        },
        "type": "object",
        "required": [
          "id",
          "format",
          "dependency_count",
          "sha256"
        ],
        "title": "SBOMUploadResponse"
      },
      "ServiceResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "project_id": {
            "type": "string",
            "format": "uuid",
            "title": "Project Id"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "project_id"
        ],
        "title": "ServiceResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "VersionChangeResponse": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "from_version": {
            "type": "string",
            "title": "From Version"
          },
          "to_version": {
            "type": "string",
            "title": "To Version"
          }
        },
        "type": "object",
        "required": [
          "name",
          "from_version",
          "to_version"
        ],
        "title": "VersionChangeResponse"
      },
      "VulnerabilityBriefResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "cve_id": {
            "type": "string",
            "title": "Cve Id"
          },
          "severity": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Severity"
          },
          "summary": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Summary"
          }
        },
        "type": "object",
        "required": [
          "id",
          "cve_id",
          "severity",
          "summary"
        ],
        "title": "VulnerabilityBriefResponse"
      },
...
    },
    "securitySchemes": {
      "APIKeyHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key"
      }
    }
  }
}